Modify DeleteDocumentCommand.cs so that Documents cant't be deleted u…

+++ REST/DMS.Application/Commands/DeleteDocumentCommand.cs
@@ -1,9 +1,9 @@
 using DMS.Application.Interfaces;
 using DMS.Domain.DomainEvents;
-using DMS.Domain.Entities;
-using DMS.Domain.Entities.Tag;
+using DMS.Domain.IRepositories;
 using DMS.Domain.ValueObjects;
 using MediatR;
+using Microsoft.Extensions.Logging;
 
 namespace DMS.Application.Commands
 {
@@ -12,23 +12,33 @@ namespace DMS.Application.Commands
     public class DeleteDocumentCommandHandler(
         IFileStorage fileStorage,
         IUnitOfWork unitOfWork,
-        IMediator mediator) : IRequestHandler<DeleteDocumentCommand>
+        IMediator mediator,
+        IDmsDocumentRepository documentRepository,
+        ILogger<DeleteDocumentCommandHandler> logger) : IRequestHandler<DeleteDocumentCommand>
     {
         public async Task<Unit> Handle(DeleteDocumentCommand request, CancellationToken cancellationToken)
         {
             try
             {
+                var document = await documentRepository.Get(request.Id);
+                
+                if (document is null)
+                    throw new Exception($"Document {request.Id} not found");
+
+                if (document.Status < ProcessingStatus.Finished)
+                {
+                    logger.LogInformation($"Document {request.Id} cannot be deleted. Is still processing.");
+                    return Unit.Value;
+                }
                 await unitOfWork.BeginTransactionAsync();
                 await unitOfWork.DmsDocumentRepository.DeleteById(request.Id);
                 await fileStorage.DeleteFileAsync(request.Id);
-                var document = await unitOfWork.DmsDocumentRepository.Get(request.Id);
                 document!.AddDomainEvent(new DeletedDocumentDomainEvent(document));
                 await unitOfWork.CommitAsync();
                 return Unit.Value;
             }
             catch (Exception e)
             {
-                await unitOfWork.RollbackAsync();
                 throw;
             }
         }
